requirements.txt
Django==3.1.4
python-dotenv==0.15.0
psycopg2==2.8.6
django-postgres-copy==2.4.2
djangorestframework==3.12.2
drf-yasg==1.20.0
